engine/pyproject.toml
[project]
name = "lego-ai-build-engine"
version = "0.1.0"
description = "AI sequential LEGO build engine"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12.3",
  "pydantic>=2.9.0",
  "pydantic-settings>=2.5.0",
  "sqlalchemy>=2.0.35",
  "httpx>=0.27.0",
  "jsonschema>=4.23.0",
  "numpy>=2.1.0",
  "pillow>=10.4.0",
  "open_clip_torch>=2.26.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0"
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["engine"]

[project.scripts]
engine = "engine.main:app"
